BacktraceReport instead of BacktraceReportBase.
