+++ Backtrace/BacktraceDatabase.cs
@@ -164,7 +164,7 @@ namespace Backtrace
         /// <summary>
         /// Add new report to BacktraceDatabase
         /// </summary>
-        public BacktraceDatabaseRecord Add(BacktraceReportBase backtraceReport, Dictionary<string, object> attributes, MiniDumpType miniDumpType = MiniDumpType.Normal)
+        public BacktraceDatabaseRecord Add(BacktraceReport backtraceReport, Dictionary<string, object> attributes, MiniDumpType miniDumpType = MiniDumpType.Normal)
         {
             if (!_enable)
             {
@@ -329,7 +329,7 @@ namespace Backtrace
         /// <param name="backtraceReport">Current report</param>
         /// <param name="miniDumpType">Generated minidump type</param>
         /// <returns>Path to minidump file</returns>
-        private string GenerateMiniDump(BacktraceReportBase backtraceReport, MiniDumpType miniDumpType)
+        private string GenerateMiniDump(BacktraceReport backtraceReport, MiniDumpType miniDumpType)
         {
             //note that every minidump file generated by app ends with .dmp extension
             //its important information if you want to clear minidump file
@@ -378,7 +378,7 @@ namespace Backtrace
                     continue;
                 }
                 BacktraceDatabaseContext.Add(record);
-                if (ValidDatabaseSize())
+                if (!ValidDatabaseSize())
                 {
                     throw new ArgumentException("Database directory has too many records or database size is bigger than in option declaration.");
                 }
@@ -391,8 +391,8 @@ namespace Backtrace
         /// <returns>False if BacktraceDatabase doesn't have more free space</returns>
         private bool ValidDatabaseSize()
         {
-            return ((BacktraceDatabaseContext.GetSize() > DatabaseSettings.MaxDatabaseSize || DatabaseSettings.MaxDatabaseSize == 0)
-                || (BacktraceDatabaseContext.GetTotalNumberOfRecords() > DatabaseSettings.MaxRecordCount || DatabaseSettings.MaxDatabaseSize == 0));
+            return ((BacktraceDatabaseContext.GetSize() < DatabaseSettings.MaxDatabaseSize || DatabaseSettings.MaxDatabaseSize == 0)
+                || (BacktraceDatabaseContext.GetTotalNumberOfRecords() < DatabaseSettings.MaxRecordCount || DatabaseSettings.MaxDatabaseSize == 0));
         }
 
         /// <summary>
