+++ Examples/Backtrace.Core/ApplicationSettings.cs
@@ -6,8 +6,8 @@ namespace Backtrace.Core
 {
     public static class ApplicationSettings
     {
-        public const string Host = @"http://yolo.sp.backtrace.io:6097/";
-        public const string Token = "328174ab5c377e2cdcb6c763ec2bbdf1f9aa5282c1f6bede693efe06a479db54";
-        public const string DatabasePath = @"C:\Users\konra\source\BacktraceDatabase";
+        public const string Host = @"https://myserver.sp.backtrace.io:6097";
+        public const string Token = "4dca18e8769d0f5d10db0d1b665e64b3d716f76bf182fbcdad5d1d8070c12db0";
+        public const string DatabasePath = "";
     }
 }
